refactor: move walker under suggestions.
