+++ packages/cspell-trie-lib/src/lib/suggestions/suggest.ts
@@ -9,7 +9,7 @@ import {
 } from './suggestCollector';
 import { TrieRoot } from '../TrieNode';
 import { clean, isWordTerminationNode } from '../trie-util';
-import { CompoundWordsMethod, hintedWalker, JOIN_SEPARATOR, WORD_SEPARATOR } from '../walker';
+import { CompoundWordsMethod, hintedWalker, JOIN_SEPARATOR, WORD_SEPARATOR } from './walker';
 
 const baseCost = 100;
 const swapCost = 75;
