+++ packages/cspell-trie-lib/src/lib/suggestions/walker/index.ts
