+++ packages/cspell-trie-lib/src/lib/walker.ts
@@ -1,0 +1,1 @@
+export * from './suggestions/walker/index';
